snake

+++ src/commands/ticket/close.js
@@ -1,4 +1,8 @@
+const db = require("quick.db")
+
 exports.run = async (client, message, args) => {
+    let prefix = db.get(`prefix_${message.guild.id}`)
+    if (prefix === null) prefix = "s."
         
         message.delete();
